+++ projects/rocprofiler-compute/src/rocprof_compute_analyze/analysis_cli.py
@@ -70,6 +70,9 @@ class cli_analysis(OmniAnalyze_Base):
                 )
 
                 pc_sampling_data = file_io.load_pc_sampling_results(path_info[0])
+                native_code_object_map = file_io.load_native_code_object_map(
+                    path_info[0]
+                )
 
                 workload.raw_pmc = file_io.process_pc_sampling_kernel_trace(
                     pc_sampling_data
@@ -91,7 +94,11 @@ class cli_analysis(OmniAnalyze_Base):
                 workload.dfs[parser.PMC_DISPATCH_INFO_TABLE_ID] = dispatch_info_df
 
                 parser.load_non_mertrics_table(
-                    workload, path_info[0], args, pc_sampling_tool_data=pc_sampling_data
+                    workload,
+                    path_info[0],
+                    args,
+                    pc_sampling_tool_data=pc_sampling_data,
+                    native_code_object_map=native_code_object_map,
                 )
                 parser.nullify_unevaluated_metric_values(workload)
                 continue